requirements.txt
pyyaml>=6.0
openai>=1.0.0
tiktoken>=0.5.0
numpy>=1.22.0
pandas>=1.3.5
matplotlib>=3.5.0
tqdm>=4.64.0
pytest>=7.0.0
colorlog>=6.7.0
python-dotenv>=1.0.0
langchain>=0.0.267
backoff>=2.2.1
tenacity>=8.2.0
requests>=2.28.0
jinja2>=3.1.2
difflib>=0.1.0
scikit-learn>=1.0.2
transformers>=4.30.0
jsonschema>=4.17.0 